+++ core/src/main/java/bio/terra/juniper/core/model/maintenance/MaintenanceModeSettings.java
@@ -1,0 +1,13 @@
+package bio.terra.juniper.core.model.maintenance;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
+@Getter @Setter @SuperBuilder
+public class MaintenanceModeSettings {
+    private String message;
+    private String password;
+    private boolean enabled;
+    private boolean disableScheduledJobs;
+}
