+++ ui-core/src/index.ts
@@ -8,7 +8,7 @@ export * from './components/forms/PagedSurveyView'
 export { SurveyAutoCompleteButton } from './components/forms/SurveyAutoCompleteButton'
 export { SurveyReviewModeButton } from './components/forms/ReviewModeButton'
 export { createAddressValidator } from './surveyjs/address-validator'
-export * from '/maintenance/MaintenanceMode'
+export { MaintenanceMode } from './maintenance/MaintenanceMode'
 
 export * from './types/forms'
 export * from './types/landingPageConfig'
@@ -20,6 +20,7 @@ export * from './types/kits'
 export * from './types/user'
 export * from './types/family'
 export * from './types/logging'
+export * from './types/maintenance'
 
 export * from './autoSaveUtils'
 export * from './reactUtils'
