+++ ui-core/src/types/maintenance.ts
@@ -1,0 +1,6 @@
+export type MaintenanceModeSettings = {
+    enabled: boolean
+    message: string
+    maintenancePassword: string
+    disableScheduledJobs: boolean
+}
